Bugfix: Project data deleted randomly
Sometimes when the SAVE event is called
the project data is empty, causing
all locally stored data to be deleted.

The problem accrued when a checked checkbox
was created. If the checkbox was meant to be
checked when loaded, i would call toggleChecked()
on the item. But this would cause the SAVE event
to fire and store an empty array.

Now i just set the isChecked variable instead.